ft-#15: refactor all codebase
