+++ src/components/Alert/index.tsx
@@ -4,7 +4,7 @@ const Alert = ()=>{
     return (
         <AlertWrapper>
         <div className="alert-container">
-
+        Copied  to clipboard successfully
         </div>
         </AlertWrapper>
     )
@@ -15,6 +15,8 @@ export default Alert
 const AlertWrapper = styled.div`
 width: 100%;
 height: auto;
+position: absolute;
+top: 0;
 .alert-container {
     width: 100%;
     height: 80px;
