+++ src/components/ChildCard/index.tsx
@@ -1,12 +1,8 @@
 import { colors } from "../../constants/colors";
 import ChildCardWrapper from "./ChildcardWrapper";
 import {FaDollarSign}  from "react-icons/fa"
+import { TTitle } from "./type";
 
-type TTitle = {
-  title: string,
-  isOpen: boolean | undefined,
-  amount: number
-}
 const ProfileDetailsTitle = ({title, isOpen, amount}: TTitle) => {
   return (
     <ChildCardWrapper>
