+++ src/components/ChildCard/type.ts
@@ -1,0 +1,8 @@
+
+type TTitle = {
+    title: string,
+    isOpen: boolean | undefined,
+    amount: number
+  }
+
+  export type {TTitle}
