+++ src/components/DetailsAction/index.tsx
@@ -1,49 +1,19 @@
-import styled from "styled-components"
 import {FaComment, FaPhoneAlt} from "react-icons/fa"
 import { colors } from "../../constants/colors"
-import { useState } from "react"
+import useCopy from "./hooks/useCopy"
+import type { TDetails } from "./type"
+import DetailsActionWrapper from "./DetailsActionWrapper"
 
 
-type TDetails = {
-    content: {
-        owner: string,
-        role?: string,
-        phoneNumber?: string
-    }
-}
-const DetailsAction = ({content: {owner,phoneNumber}}: TDetails)=>{
 
-console.log(phoneNumber)
-    const [isCopied, setIsCopied] = useState(false);
+const DetailsAction = ({content: {owner,phoneNumber,setIsCopied}}: TDetails)=>{
 
-    async function copyTextToClipboard(text: string) {
-      if ('clipboard' in navigator) {
-        return await navigator.clipboard.writeText(text);
-      } else {
-        return document.execCommand('copy', true, text);
-      }
-    }
-
-
-    const handleCopyClick = () => {
-        if(phoneNumber)
-        
-        copyTextToClipboard(phoneNumber)
-          .then(() => {
-            setIsCopied(true);
-            console.log("copied")
-            setTimeout(() => {
-              setIsCopied(false);
-            }, 1500);
-          })
-          .catch((err) => {
-            console.log(err);
-          });
-      }
+const {handleCopyClick} = useCopy(phoneNumber, setIsCopied)
 
     return (
         <DetailsActionWrapper>
             <div className="action-container">
+                
                 <div className="info-cont">
                 <div className="profile"></div>
             <div className="name">
@@ -57,8 +27,10 @@ console.log(phoneNumber)
             
 
             <div className="action-button">
+                <div className="phone-container">
                 <input className="phone" value={phoneNumber} onClick={handleCopyClick} readOnly />
-                <FaPhoneAlt  size="20px" color="red"/>
+                <FaPhoneAlt className="phone-icon" size="20px" color="red"/>
+                </div>
                 
 
                     <div className="comment">
@@ -74,72 +46,3 @@ console.log(phoneNumber)
 
 export default DetailsAction
 
-const DetailsActionWrapper = styled.div`
-    width: 100%;
-    height: 60px;
-    display: flex;
-    justify-content: center;
-    margin-top: 20px;
-
-    .action-container {
-        width: 90%;
-        height: 100%;
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-
-        .info-cont {
-            display: flex;
-            align-items: center;
-            
-        }
-
-        .profile{ 
-            width: 50px;
-            height: 50px;
-            border-radius: 50%;
-            background-image: url("https://i.pinimg.com/564x/99/03/34/9903345b0d415013d2a840842451e6d0.jpg");
-            background-size: contain;
-        }
-
-        .name {
-            padding: 8px;
-
-            .user-name{
-                font-size: 16px;
-                font-weight: bolder;
-            }
-            .user-role {
-                font-size: 14px;
-            }
-        }
-
-        .action-button {
-            display: flex;
-            justify-content: space-between;
-            padding: 12px;
-
-            .phone {
-                width: 40px;
-                margin-right: 10px;
-                background-color: ${colors.primary.red}; 
-                display: flex;
-                justify-content: center;
-                padding: 10px;
-                border-radius: 10px;
-            }
-
-            .comment {
-                width: 40px;
-                margin-right: 10px;
-                background-color: ${colors.primary.lightCyan}; 
-                display: flex;
-                justify-content: center;
-                padding: 10px;
-                border-radius: 10px;
-            }
-
-
-        }
-    }
-`
