+++ src/components/DetailsAction/type.ts
@@ -1,0 +1,10 @@
+type TDetails = {
+    content: {
+        owner: string,
+        role?: string,
+        phoneNumber?: string,
+        setIsCopied:  React.Dispatch<React.SetStateAction<boolean>>
+    }
+}
+
+export type {TDetails}
