+++ src/components/DetailsDescription/DescriptionWrapper.ts
@@ -1,0 +1,41 @@
+import styled from "styled-components";
+import { colors } from "../../constants/colors";
+
+const DetailsDescriptionWrapper = styled.div<{isExpand: boolean}>`
+width: 100%;
+display: flex;
+justify-content: center;
+margin-top: 20px;
+
+    .details-container {
+        width: 90%;
+        height: ${(props)=> props.isExpand ? "auto": "40px" };
+        overflow-y: hidden ;
+        position: relative;
+        font-size: 14px;
+        font-weight: 600;
+        overflow-x: hidden;
+
+        .expand {
+            position: absolute;
+            margin-left: 20px;
+            bottom: 0;
+            margin-top: 10px;
+            color: ${colors.primary.cyan};
+            font-weight: bolder;
+            z-index: 20;
+            background-color: ${colors.primary.white};
+            font-weight: bolder;
+            width: 100%;
+            padding-left: 10px;
+        }
+
+        span:first-child {
+            z-index: 0;
+        }
+
+     
+    }
+`
+
+export default DetailsDescriptionWrapper
