+++ src/components/DetailsDescription/type.ts
@@ -1,0 +1,6 @@
+
+type TDescription  = {
+    desc: string;
+}
+
+export type {TDescription}
