+++ src/components/Header/type.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from "react"
+
+type THeader = {
+    children: [ReactNode,ReactNode]
+}
+
+export type {THeader}
