+++ src/components/Map/GoogleMapWrapper.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components"
+
+const GoogleMapWrapper = styled.div`
+    width: 100%;
+    height: 200px;
+    display: flex;
+    justify-content: center;
+    margin-top: 10px;
+
+    .map-container {
+        width: 90%;
+        border-radius: 40px;
+        overflow: hidden;
+    }
+`
+
+export default GoogleMapWrapper
