+++ src/components/Verification/type.ts
@@ -1,0 +1,6 @@
+type TVerification = {
+    rating: number,
+    isVerified?: string
+  }
+
+  export type {TVerification}
