+++ src/pages/Details/DetailsWrapper.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components"
+
+const DetailsWrapper = styled.div`
+    width: 100%;
+    border: 1px solid red;
+    display: flex;
+    flex-direction: column;
+    z-index: -1;
+    overflow-y: hidden;
+    height: 100vh;
+    position: relative;
+`
+
+export default  DetailsWrapper
